test: move all e2e tests into a single package
