+++ .github/workflows/ci.yaml
@@ -97,5 +97,5 @@ jobs:
         env: 
           DOCKER_HOST: ${{ matrix.docker_host }}
         run: |
-          docker build -t topo-e2e-target:latest ./internal/testutil/test-container
+          docker build -t topo-e2e-target:latest ./e2e/testdata/test-container
           go test ./... -v -covermode=atomic -coverprofile=coverage.out
