+++ internal/deploy/docker/deploy_test.go
@@ -3,7 +3,6 @@ package docker_test
 import (
 	"bytes"
 	"fmt"
-	"os"
 	"path/filepath"
 	"testing"
 
@@ -136,41 +135,6 @@ func TestNewDeployment(t *testing.T) {
 
 func TestDeployment(t *testing.T) {
 	testutil.RequireDocker(t)
-
-	t.Run("Run", func(t *testing.T) {
-		target := testutil.StartTargetContainer(t)
-
-		t.Run("builds images, transfers them, and starts services", func(t *testing.T) {
-			remoteDockerHost := ssh.Host(target.SSHDestination)
-			tmpDir := t.TempDir()
-			dockerFilePath := filepath.Join(tmpDir, "Dockerfile")
-			dockerFileContent := `
-FROM alpine:latest
-CMD ["tail", "-f", "/dev/null"]
-`
-			testutil.RequireWriteFile(t, dockerFilePath, dockerFileContent)
-			composeFilePath := filepath.Join(tmpDir, "compose.yaml")
-			composeFileContent := fmt.Sprintf(`
-name: %s
-services:
-  busybox:
-    image: busybox
-    command: ["tail", "-f", "/dev/null"]
-  a-service:
-    build: .
-`, testutil.TestProjectName(t))
-			testutil.RequireWriteFile(t, composeFilePath, composeFileContent)
-			t.Cleanup(func() { testutil.ForceComposeDown(t, composeFilePath) })
-
-			deployOpts := docker.DeployOptions{TargetHost: remoteDockerHost}
-			d, _ := docker.NewDeployment(composeFilePath, deployOpts)
-
-			err := d.Run(os.Stdout)
-
-			require.NoError(t, err)
-			testutil.AssertContainersRunning(t, remoteDockerHost, composeFilePath)
-		})
-	})
 
 	t.Run("DryRun", func(t *testing.T) {
 		t.Run("prints all commands", func(t *testing.T) {
